[vim] Increased support for Haskell development
This commit was focused around my new interest in the Haskell
programming language. There where a lot of plugin testing, but I believe
that the current selection is lightweight and highly usable.

I also removed a lot of plugins that I don't know how to use. Some of
these might be added later, once I get time to take a look at them. Some
of the markdown plugins where replaced by vim-pandoc. Everyone might not
support this, but there is no need to have more plugins doing the same
thing.